Moving the single top model (KappaTKappaV)
Moved the model `KappaTKappaV` outside of the `LHCHCGModels.py` as it is
neither compatible with the other couplings models (includes an overall
scaling parameter "r") and ignores VH production processes

Model is now in a separate file `SingleTopModels.py`
